Use gfortran rather than lsp as a Fortran flycheck checker
LSP was flagging things that weren't really errors - in particular:
   character(len=*) :: something
was being flagged as 'len' not defined (or something like that).

And LSP was *not* catching true problems, like a true undefined
variable.

The gfortran checker isn't working perfectly either, but I think I'll
have a better chance getting it to work.

But actually, this syntax doesn't seem to be working quite right yet:
it's still using lsp as a checker.