Improve flycheck setup
- Stop running flycheck when inserting new lines and idle: I was finding
that a bunch of errors would appear temporarily, which was
annoying. (For example, as soon as I insert a new line in python, I get
a pylint error due to trailing white space; this lasted until I typed
something and then was idle for 5 seconds). I notice that VSCode only
checks on save, so that seems like a reasonable option. It could also
work to go back to the default idle delay (0.5 sec) so that the errors
go away quickly as I type, but that causes the mode line to flash a lot
as flycheck constantly runs, so if I did that, I would want to try to
find a way to prevent that frequent updating of the mode line.

- Use pylint instead of lsp for python linting