Disable lsp-ui-sideline
I think I prefer seeing flycheck diagnostics in the minibuffer rather
than in the "sideline". This disables other things in the sideline,
too. I haven't played with this enough to know if I necessarily want to
disable all of it, but my initial feeling is that the sideline stuff is
too unpredictable in its location (since it tries to find a close line
above the current line with enough whitespace... which may end up being
far away or nowhere).