Use pyright as python LSP server
I've seen recommendations for this over the other options (pyls and the
older Microsoft python language server): apparently this is Microsoft's
latest version.

Compared with pyls, what I like more is that the description of a
variable (when the cursor is on a variable) just gives that variable's
type, rather than giving extraneous information (pyls gave something
like how to construct variables of that type).

To get this to work, I had to do:

brew install npm
sudo npm install -g pyright