Revert "Change how lsp mode is loaded with python"
This reverts commit 68f97a97dcf3648062f75913efa426fb16049131.

I'm going to do this with a global setting for now: Doing it via
dir-locals would require either committing this file (which could cause
problems for other people using emacs – especially due to the question
about loading unsafe variables), or would require me to add this to my
.gitignore, either in the repo or globally... and I don't really want to
add it globally. dir-locals may be a good solution eventually, but I
think I'll go with the simple solution of doing it globally for now.