Get rid of lsp's eldoc on hover, introduce a func that does it better
Now I can run A-d to bring up the function documentation in a little
window below, then A-v to delete that window. If there is too much to
fit, I can go down to that window and split it with C-x 2 so that I can
see both the argument list and individual arguments at the same time.