Switch the prefix used for a lot of key bindings, and other tweaks
I feel like it makes more sense to use A for editing commands and things
I want to repeat multiple times in succession. So I'm moving a lot of my
commands that I had on A to H instead.

I was also finding it unintuitive to have super on both sides of the
keyboard, so I'm swapping where I have super vs. hyper.

I also added, deleted or changed a few other key bindings.