Stop using company-sort-by-occurrence
At least in python, it seems like the completion I want is more often
first in the list without this. In particular, it seems like the fuzzy
matching prioritizes non-fuzzy results if I do NOT include this
transformer, but with this transformer, fuzzy results often show up
first. I haven't done any testing in Fortran without this yet.