Add company setting
The main motivation is: I was finding that, if I started a manual
completion, then deleted all the text I started with using
ctrl-backspace, LSP completion would sometimes temporarily get broken in
that buffer. This setting seems to fix that issue. In addition, though,
it just seems like the way things should work.