Actually, change company completion to be manual only
From a bit of trial, I don't love having it pop up after a delay: it
kind of makes me anxious that it's going to pop up if I wait too long to
type. I think that was what led me to make the delay 0 when I first
started using company mode. But now that (1) I have changed the company
key binding to be really easy to press, and (2) my workflow is different
due to the use of LSP (I'm not finding myself repeatedly tab-completing
the common prefix anymore, partly because of the fuzzy matching making
that not work well; I'm more often getting some prefix then selecting
the right choice): it actually seems to work well to require manual
invocation of completion.