Go back to having company-idle-delay set to 0 by default
From some further experimentation, I am not so sure that this is really
what's responsible for the slowness in completion in case.py. But even
if it is, I realized that a better solution is to set it to 0 by default
but add a function to toggle it for when I'm dealing with a buffer where
it's an issue - because there really are some benefits to having it
trigger immediately.