Bind M-SPC to company-complete-selection
With fuzzy matching, I'm not finding it too useful to complete a common
prefix. So I more commonly want to complete with the current selection,
so make that easier.

I'm not positive I like this: it's nice that it's easier, but it feels
less symmetrical with the operation of M-SPC before company has been
started. So I may decide to change this back.

But one rationale for this is: A-M-SPC now invokes company with dabbrev,
and that's where I am more likely to want to use
company-complete-common, so it makes sense for A-M-SPC to also do
company-complete-common once it is invoked.

Also add some comments.