Change company minimum prefix length to 1
The issues that originally led me to setting this to 4 seem to only
apply with the dabbrev-based completion, not the smarter lsp/capf-based
completion. My old notes (from 2019-07-03) about this were:

    I'm changing company minimum prefix length from 3 to 4, partly to avoid
    company from kicking in for "end", and (the real reason I'm changing
    this now) partly because 3 letter prefixes seem not to show up as whole
    word options in the list of completions (i.e., if there is something
    like abc and another thing like abcdef, then when completion kicks in
    after typing abc, just keeping abc isn't an option! This seems to be a
    bug, and I'm going to avoid it by having it not even try for completion
    at 3 letters, since I was thinking of making that change anyway for the
    sake of having it try to complete "end").

but these issues don't exist anymore, at least in the Fortran file I
tested with.

Note, however, that I have not actually tried using this minimum prefix
length of 1; I may decide that I don't like it.