Stop using unicode character for org-ellipsis
Even with the setting of org-cycle-separator-lines to 1, I'm still
having issues when doing org-show-todo-tree: this is sometimes shown as
a hollow box instead of the desired character.

So for now going back to using standard ascii characters, along with
some background highlighting to make this more visible. If the bug in
the unicode ellipsis is fixed in org mode, then I'd imagine reverting
this commit.