Use show-paren-mode instead of show-smartparens-mode
It seems like some laggy behavior is coming from show-smartparens-mode -
at least when dealing with strings (for which this mode highlights the
matching quotes), and *maybe* in other cases. Other than matching the
quotes, I'm not sure this provides much beyond show-paren-mode, so I
think I'm going to use the built-in show-paren-mode. Indeed
https://github.com/Fuco1/smartparens/wiki/Show-smartparens-mode
acknowledges that show-smartparens-mode can be slower.