Hide org emphasis markers
The main motivation is to hide the distracting ~ markers in inline
code (these make things hard to read sometimes). I have mixed feelings
on whether I like to show or hide emphasis markers in general: hiding
them makes the text look nicer when reading, but can make it harder to
edit. I'll give this a try and see if I like it.