Visual-fill-column: keep fringes inside margins
The main motivation for making this change was: sometimes when I reach
the fill column in visual-line-mode / visual-fill-column-mode, the
cursor temporarily jumps to the very end of the line. I find this
distracting. From a brief trial, it seems like this setting may fix the
problem. Also, it's kind of nice having this visual indicator of when
the line is going to wrap.