Use Shift-Return to recenter
The main motivation for this key binding is to have an easier way to
recenter during isearch, which is something I frequently want to do.
This shift-return keybinding works well when I'm going through isearch
with the arrow keys (which I have customized to go through the isearch
results). So I initially thought I'd just do this binding in the isearch
mode map. However, I think I'll try just doing it globally: it might be
nice in other contexts, too.