Turn off auto import completions for pyright
There are pros and cons of this, but I think the cons of auto import
completions (longer list making it harder to do quick completions,
slower performance) may outweigh the pros (candidates that aren't
currently available, automatically writes the import statement). I at
least want to try with it off for a little while.