LSP: turn off count of code-actions on the modeline
I discovered that this was responsible for 100% CPU usage and hangs of
the jedi language server: it seemed that, with every move of the cursor,
LSP was asking the server for the actions at the current point, and this
request was taking a while to return.

Furthermore, this count seems fairly useless. So I'm turning it off.