Make flycheck info issues much less visible
The original motivation for this is not liking the constant flagging of
trailing whitespace when running pylint. But this could be more general:
I think info should be a light gray (both in the squiggle and in the
margin) so it isn't too distracting as I type.

I'm not sure yet whether I like the idea of setting
flycheck-color-mode-line-info-face to the same as
flycheck-color-mode-line-success-face (gray50): My point in doing this
was to avoid colors flashing due to addition and removal of trailing
whitespace in a python buffer checked by pylint. But I may decide to
make the info-face darker gray, or possibly make both of them darker
gray. (I wanted to make the flashing during checking less visible when
in a buffer with no current errors, but having the text a bit darker
than it currently is would probably still be fine.)