Map A-g to repeat, and change some A mappings to s
I learned about the command to repeat the last command. With this, it's
less important for some things to be under A bindings in order to be
easily repeatable: I can do a more awkward keybinding then repeat it
with A-g. So I'm changing some bindings (particularly for flycheck and
winner) to more mnemonic s- bindings.