Use Page Up / Page Down for my scroll-up / scroll-down commands
This feels *slightly* more ergonomic (at least for the version by 3,
which now uses Meta instead of Alt; and at least no worse for the A-M
versions, and I think a bit better) and *slightly* easier to remember
than the arrow key versions. In addition, it frees up the arrow keys for
other future key combos, if desired.