Move by logical lines rather than visual lines
The original motivation for these changes was a problem in org-mode, as
noted in the comment:

    I'm having problems with the combination of visual-line-mode and
    org-mode (at least when there are some hidden sections; I'm not sure
    if there are problems if there are no hidden sections): line
    movement commands (up and down arrows, C-a) sometimes cause the
    cursor to jump to some far-away line. Setting line-move-visual to
    nil (undoing its setting to t by visual-line-mode) seems to solve
    the up/down arrow key issue, but not the C-a issue. (Oddly, running
    org-beginning-of-line (which is what C-a is bound to) doesn't
    exhibit this problem, but I can't figure out how to solve the
    problem for C-a.) So for now, I'm working around this issue by
    avoiding visual-line-mode, and instead turning on some individual
    pieces of it. Arguably, line-move-visual would be a good thing when
    using visual-line-mode (or my variant on it); but I'm going to try
    having it off in order to avoid this issue in org mode; this means
    that line movement will happen by logical lines rather than visual
    lines. And actually, arguably that is actually a better behavior
    anyway.

As noted there, I'm solving the org-mode problem by avoiding turning on
visual-line-mode directly, but instead doing the pieces of it I want
myself - and *not* doing the pieces I don't want (changing
line-move-visual and the key-bindings; I'm also not bothering with
setting truncate-partial-width-windows to nil because I don't think I
need it).

But in doing this, I also decided to set line-move-visual to nil: I feel
like this is probably better behavior for editing code, regardless of
what I end up doing for org mode and similar non-code modes.