Revert "Move by logical lines rather than visual lines"
This reverts *most of* the changes in commit
794f630ffe2fe5c71859f27bf5b44ca625b55230.

However, I'm leaving in place the changed behavior for non-document-mode
things (e.g., code), because I think I'll prefer that.

Why I'm reverting this: I found that, even with these changes, there
were still issues with right arrow or C-n. It seems the problem was due
to org-hide-emphasis-markers. I'm going to fix this more robustly in a
following commit.

I was a bit on the fence about whether to go back to visual-line-mode or
stick with my partial-visual-line-mode that keeps movement-by-lines
working on logical lines. But I think I'll go back to the full
visual-line-mode: while there are pros & cons, I think it's a bit
better, given that I will fix this issue with line movement causing the
point to move to a crazy location.