Use org-appear-mode
This makes emphasis markers appear and disappear dynamically. The main
motivation is that this seems to truly fix the issue where, when I move
the cursor around text with invisible markers, sometimes the cursor
would jump to some faraway location.

I was worried that it would be annoying / distracting to have these
markers appear and disappear, shifting other things around. But from a
quick trial it isn't too bad, and overall I think I like
org-appear-mode. But if I end up being bothered by the shifting of text,
I'll stop using org-appear and turn off org-hide-emphasis-markers to
avoid the issue of the cursor jumping to faraway locations.