Edit-server: go back to NOT using a new frame
I was finding the use of a new frame annoying, both because: (1) I'd
prefer to have all of my other open buffers available alongside the
edit-server window, and more importantly (2) with the use of a new
frame, whenever you save, the frame is deleted and recreated; this means
that, if I had split the frame, opened any other buffers in that frame,
etc., all of that work is lost (and winner doesn't seem to be able to
recover it, maybe because that frame is totally gone).

The problem with NOT using a new frame is that when you're done (with
C-c C-c), the current window gets deleted (see also
https://github.com/stsquad/emacs_chrome/issues/172). I'll work around
that by using winner-undo.

I thought about adding this (in misc.el; not yet tested):

(require 'edit-server)
;; This is a total hack to prevent the edit-server from deleting the current window when
;; finishing. Otherwise, when edit-server-new-frame is nil, when finished editing, the
;; edit server's window is deleted, which I don't like. Ideally the edit-server code
;; wouldn't do that, but this function is a hacky workaround to prevent it.
(defun my-edit-server-done ()
  (interactive)
  ;; Here is the total hack: we temporarily set edit-server-new-frame to t to trick
  ;; edit-server-done into not calling delete-window. This works for now because the only
  ;; use of edit-server-new-frame is in a conditional to decide whether to call
  ;; delete-window, but if it were used for anything else, this might break.
  (let ((edit-server-new-frame t))
    (edit-server-done)))

but I'll just add that if winner-undo isn't enough of a fix.