Swap keybindings for forward-symbol and forward-to-whitespace (& back)
I think this will be easier to remember based on the position of my
modifiers: My thumb moves a little for M, a little more for M-A, and a
little more still for A alone. Similarly, the cursor will move a little
for M, a little more for M-A, and a little more still for A alone.