Go back to org-cycle-separator-lines = 1 after all
I like that this doesn't require me to do an extra command when viewing
an outline of the document. This is a little more spaced out than I'd
ideally like, but not by *too* much.

Part of the motivation for returning to this setting is: I realized that
my new key binding for toggling spacing (H-2) doesn't work in org-goto
buffers. So it only works when I'm viewing an outline via cycling
visibility or via org sparse tree. So I'm optimizing more for the
cycling-visibility case rather than trying to keep all of these things
consistent in how I get to double spacing.