Turn on org-indent-mode, and make some changes for it to look good
The initial motivation for turning on org-indent-mode was being able to
see the outline structure more easily at a glance when I'm just showing
headings, viewing a sparse tree, or in org-goto. But I'm thinking that
it might have benefits more generally - giving me a better sense of
where I am in the outline structure while writing text.

However, there were some problems that I had to fix:
- Wrapped lines sometimes went to the far left; turning off
adaptive-wrap-prefix-mode seems to fix this. I may want to leave it on
for non-org modes, but for now I'll try just turning it off always and
I'll see if that causes any issues.
- Because I use variable-pitch-mode, things weren't lining up nicely.
Making org-hide fixed-pitch makes this better (not 100%, but better than
before); note that this works because org-indent inherits from
org-hide. (See also
https://emacs.stackexchange.com/questions/7429/how-to-customize-org-mode-indentation)