Disable show-paren-mode in org-mode
This is to work around the fact that show-paren-mode has issues with
org-indent-mode: lines sometimes jump to the left margin when
parentheses are being matched. show-paren-mode is apparently a global
mode; it isn't enabled initially in org-mode, but if a programming-mode
buffer has already been opened that enables it, then it gets enabled
globally, including in the org mode buffers. So disable show-paren-mode
in org mode to work around this issue.

If I stop using org-indent-mode, then I should back out this change. And
if this problem gets fixed, I should also back out this change - since
it could be nice to have parentheses matching in org mode.