Go back to default delay for diff-hl-flydiff
I like it better when the updates to the diff-hl indicators are
near-instantaneous: with the delay, the change to the indicators a
couple of seconds after I make a change could be distracting (at least
when I'm paying attention to it).

The original reason for the change from the default (0.3 sec) to 2 sec
was in the hopes that this change would improve lagginess issues (see
532701259b62bdfe7c6f784cc32292f8b9e309b0). I can't remember if it
actually did. I should keep an eye on whether this change increases
laginess issues.