Org: do not insert blank line before new heading
I'm not sure if I'll like this or not, but want to try it. The
motivation is: If I'm making a bunch of little todo items, I sometimes
don't want a blank line between them. The main reason for that is if I'm
showing a collapsed outline, I don't want an ellipses if there is no
real content under a heading (and all that's under a heading is a blank
line).