Turn off most pyright type checking
I'm changing from typechecking mode of basic to off, because in cime I'm
getting too many errors due to questionable usages (for example, when a
variable is initialized to None but really has some type associated with
it – then you try to do something like call a list operation on the
variable and pyright says you can't because its type may be None) – and
these false errors were distracting.

I'll revisit this if we start being more disciplined about typing in our
code.

It would also be possible to configure this on a per-project basis,
and/or turn off just certain issues. See
https://github.com/microsoft/pyright/blob/main/docs/configuration.md#diagnostic-rule-defaults.
The main thing I'd want to turn off is reportGeneralTypeIssues. One
thing I might want to keep on is reportUnboundVariable.