Org: don't insert a blank before a new entry
Motivation: to keep my spacing more consistent: never have a blank line
after a heading. If there is content under a heading, then can have a
blank line after the end of the content (though this wouldn't be
necessary). This would mean that the only heading with ellipses would be
ones with content under them (this was a key motivation for this rule).
It would also mean (for better or for worse) that, in outline view, no
headings have blank lines between them.

To accomplish this consistently, I'm setting org-blank-before-new-entry
to nil for headings. When I do want a blank line, I'll use A-M-o to
insert a blank line after I have created the heading. Note that I have
created a new function for this, similar to crux-smart-open-line-above
but without putting the point on the new line above.

(The suggestion here could possibly help do this automatically:
https://emacs.stackexchange.com/questions/13311/make-org-blank-before-new-entry-distinguish-between-a-todo-list-and-a-text-outli)