Company: make A-M-SPC complete with the current selection
Make both M-SPC and A-M-SPC run company-complete-selection. This feels
easiest because I sometimes trigger company with M-SPC and sometimes
with A-M-SPC (the latter for completing with dabbrev); when triggering
with the latter, I naturally want to press the same thing to complete
with the current selection because my fingers are already there.

Then make A-SPC call company-complete-common. (I'm not sure if I'll ever
use this.)