Turn off flycheck for Fortran (f90-mode)
To get the gfortran checker to work, I would need to have a build
directory sitting around that I could point to for includes. This could
be a good long-term solution, but it will take some work to get this
working smoothly. Another possible need would be to do something like
what VSCode's https://github.com/krvajal/vscode-fortran-support does:
updating the module files whenever a file is saved; or adding some
feature to manually update the module file for the current file so that
diagnostics for files using that file are correct (e.g., if I add a
function or change arguments of a function, then I'd want files using
that function to have updated diagnostics; though I could also achieve
that by rebuilding the whole thing).

The LSP checker could be easier to get up and running, but currently is
limited in what it checks, falsely flags 'character(len=*)' as an error
for some code, and only updates on save. So the gfortran checker seems
like probably the better option.