Org: Do not show subtasks in agenda
I'm not sure whether I'll want this or not, but at first glance, I
prefer not to have subtasks listed.

If I keep this, I may want to undo the recent customization where I
added indentation to the agenda view based on the level: although this
shouldn't hurt, it is probably unnecessary with this new customization.