Use s-g instead of s-o for org-mode things
I was finding s-o to be a painful reach for something I was using so
often. s-g is more ergonomic.

s-g had been bound to isearch-repeat-forward, but that doesn't seem very
useful outside of the general isearch, and even if I kept it, it
wouldn't make sense to have it on the super key with my keybinding
strategy.