Org: For headline fonts, use CharterSpacedAbove
This is a new version of the Charter font that I made with a version of
the font-line tool that I have modified to allow just changing the
ascent spacing, not the descent spacing. With this version of the font,
I feel I can skip having a blank line after content before the next
heading. I like this both because it means I no longer need blank lines
before some headings (which I was finding was a bit hard to maintain),
and because I like the look and feel a bit better than having blank
lines before some headings.