Org: add a background for the PR-# states
The initial motivation was making PR-1 more visible when on a line with
a dark red headline (which is a typical place for it to appear). But I
realized this is also generally useful in making the project headings
stand out in a big org document.

I'm using a color that contrasts with the PR-1 color, to make PR-1 even
more visible.