Org: add properties for the new "large" tag
I'm using a tag of "large" to denote projects that have their own
subprojects. The original motivation for this is helping me know which
projects I need to dive into in doing my weekly planning (it's not
enough to mark "water isotopes" as pr-1: this is a large project with
its own big subprojects, so I need to go in there and prioritize the
subprojects).

This commit adds some properties for this tag.