Org: change org-agenda-dimmed-todo-face
The default is gray. This is used when a task is "blocked" simply
because it has children. It does seem helpful to denote this in some
way, but the gray is too extreme: it makes it hard to read the text, and
it masks the nice coloring of the various todo states. So I'm going with
a more subtle italic rather than graying it out.