Org: Use bold font for headlines
Original rationale: make headlines stand out more, particularly so it's
easier to see at a glance when one section ends and the next begins. I
originally was thinking of this for notes, but it could apply to a todo
list as well. This is particularly to aid glanceability now that I don't
have a blank line between the content of one section and the next
heading.

But actually, I think I find this a bit more readable for a todo list,
even when everything is collapsed so I'm just looking at the headings.
I'm not sure about this, though: I'll try it out for a bit and see how I
like it: If I find I don't like it for a todo list, then I'll back it
out, since the look & feel of a todo list is more important than for
notes.