Org: underline todo states, strike-through CANC
Underlining the todo states makes them feel more todo-like, and a bit
more visually distinct. It may be that this wouldn't be as valuable with
non-bold headlines (when the todo keywords were bold but the rest of the
headlines weren't), but I think I like it now. (I tried having the
underline be the color of the text, as well as lighter or darker gray; I
like this fairly light gray that is visible but not too
eye-catching.) (I also tried a box around the text as an alternative to
the underline, but felt that was too visually distracting.)

However, I'm *not* underlining done states (accomplished by having them
inherit from org-done face, which doesn't have the underlining).

I am, however, adding a strike-through to CANC to make it more visually
distinguishable from DONE.