Org: make drawers less visible
The motivation for this is making the PROPERTIES drawer less visible (so
that if a section has a VISIBILITY property, for example, then the
drawer doesn't stand out too much). This is accomplished with the
org-drawer face. But it also seemed like the contents of the properties
drawer could be less visible (I was trying to make the ellipsis after a
folded drawer be less visible; I'm not sure if I managed to accomplish
that, but it seems like it's fine and good the contents of a properties
drawer to be less visible anyway); this is accomplished with
org-property-value and org-special-keyword face.

Making org-special-keyword face less visible also impacts the CLOSED
keyword; I think that's actually good.