Change org ellipsis
The original motivation for this was to make the ellipsis less
eye-catching and distracting on a properties drawer. But I realized that
I've generally been unhappy with my current ellipsis, especially now
that I have taller headlines, so the background color extends for a big
height, which I find somewhat distracting. Furthermore, now that I don't
have blank lines between headers and content, it should be safe to use a
unicode character (avoiding the weird block issue I experienced a few
months ago that led me to abandon this).

Specifically:

(1) Stop setting the background color for org ellipsis. I feel like this
is an improvement over what I had before, even without the change
in (2): I don't need the extra visibility of the tall background
color (tall because it extends the whole vertical height of the line).
Interestingly, there still seems to be a background for the ellipsis,
but now it is a lighter color and just as tall as the ellipsis, not
extending the whole line (I can't tell where that's coming from).

(2) Switch back to a unicode down arrow. I like the appearance of this
more than the triple-dots: it does a better job at intuitively
indicating that there is extra content. It is maybe slightly more
visible than the triple dots, for better or for worse, but in any case
it seems more visually appealing.

With this new setting, the ellipsis character is small at the end of the
property drawer lines, so it isn't too visually distracting.

Compared to what I had before, I find this new ellipsis slightly to
somewhat less stand out-ish, which I think is a good thing. If anything,
I might still tweak things a bit more to make the ellipsis stand out
even less (at least slightly)....