Use org-disputed-keys to handle my org mode key changes
This should be nearly equivalent, but seems like the "right" way to do
it.

I do see that it leaves M-Ret and M-S-Ret defined (along with the new
M-A-Ret and M-A-S-Ret), but I guess that's okay.

Notes about these key bindings (same as before): I am replacing uses of
M with M-A. The main motivation is because I use M-arrows for other
purposes. But I wanted to remain consistent, especially for M-S-arrow:
it feels better and easier to remember and press if those are replaced
by M-A-S-arrow. I am then also doing the same for M-Ret and M-S-Ret for
consistency, and because (for example) it makes it easier to insert a
heading then immediately promote / demote it.