Use ctrl for windmove
This feels (significantly?) easier than hyper, and I can do it even when
my palms are lazily resting on the keyboard.

This shadows ctrl-left/right for backward/forward word, but that does
essentially the same thing as M-b/M-f, and it doesn't seem necessary to
have two different bindings (that are approximately equally ergonomic)
for essentially the same thing.

Then also change my move-text bindings to M-A instead of ctrl up/down. I
actually like this better. This is shadowed by org-specific bindings in
org mode, but I actually think this is a good thing: it's often a
mistake to try to move a single line in org mode, and conceptually M-A
up/down now does the same thing - just in org mode, it moves a whole
subtree up/down.