Org: Go back to somewhat more visibility of some elements
I think the original motivation for this was making deadlines more
visible - by making org-special-keyword and/or org-date more visible. I
like the compromise of having these be a color that is more visible but
smaller text. This does make CLOSED stand out more than it needs to, but
I don't think that's a huge deal: I won't typically have closed things
expanded.

I'm also making org-property-value be family Hack so that it matches the
face of org-special-keyword. (I might be able to do this in an
alternative way by making org-variable-pitch mode turn
org-property-value into fixed pitch; I haven't investigated this
option.)

Note that I'm still maintaining less visibility of org drawers.

Note: I experimented with making the insides of a drawer (specifically,
org-property-value and org-special-keyword) bigger, to see what that
does to the size of the ellipsis on the drawer. It seems that the
ellipsis size just depends on the org-drawer face, not what's inside the
drawer. So I could safely do that without making the ellipsis too
big. (Previously, when the ellipsis was ... with a background color, I
thought that *maybe* the size - and I think in particular, the height of
the background color - *may* have depended partly on the size of the
text inside the drawer. Now that doesn't seem to be the case - and I'm
not sure if it was even the case before.)