Use standard org bindings for meta
Stop using meta-alt (M-A) in place of meta (M) in org mode: go back to
standard org mode bindings.

I'm not positive that I'll stick with this, but I want to give it a try.

The main motivation for this was that I wanted to reconcile my recent
org C-M arrow bindings with my M-A org bindings: I'd like my
recently-added org arrow bindings to use the same left hand modifiers as
the other org meta bindings, so that I can navigate to a heading then do
something like move it up or down or insert a new heading. So I was
originally thinking that I should use C-M-A for those. (And that feels
about as easy as C-M, even when I add the shift key, so that seems
fine). But to do that, I'd need to change divvy to use C-M, or something
else.

I considered changing the divvy bindings, but C-M-arrow is already taken
by some bindings in emacs, so I'd either need to change those things or
use something different for divvy.

So I decided to try using the standard org mode bindings. Now that I'm
using org heavily, this kind of makes sense anyway, since it will
probably make various things easier in org mode. So I didn't think too
much more about changing the divvy bindings, since I started to feel
like I'd like to try switching to using the standard org mode bindings,
and avoiding having conflicts with those.

To do this, I needed to remap my existing M-arrow bindings: Use
M-A-arrows where I currently use M-arrows, and use M-arrows where I
currently use M-A-arrows. The M-A bindings are a little
harder (especially noticeable for forward-paragraph and
backward-paragraph), but not too much so, and may be worth it for the
sake of using standard org bindings. (And M-A-arrows are still a bit
easier than the default binding for forward/backward paragraph of
C-arrows, on my Kinesis layout.)

Update after a bit of experience with this: The new M-A-arrow bindings
for things where I used to use M-arrow are definitely somewhat harder.
I'm especially noticing it for previous-buffer and next-buffer. But I'm
continuing to try this, to see if the easier org mode bindings make up
for this.