More ergonomic binding for my-pop-local-mark-ring
Use C-, for my-pop-local-mark-rink (equivalent of C-u C-SPC). This is
inspired by the use of M-, for xref-pop-marker-stack.

This requires me to change my other-frame binding from C-, to C-' (which
seems approximately equally as good). I have made the same change at the
system level, for the system-level shortcut for cycling between windows.