Stop using esc, home and end for hyper and super
I'm finding that I like using M-h and C-h for hyper and super. So I no
longer have a need to use esc, home and end for this purpose. Because I
don't need esc anymore, I have put a tab key where esc used to be (top
left of my R thumb). The change in this commit means that the home and
end keys are pretty useless in emacs, but I feel like that's okay: it
seems good not to have the temptation to use these hard-to-reach keys.