Use home and end for scrolling by 3
Because I'm no longer using home and end for hyper and super, and
because the built-in functionality of home and end (going to the
beginning and end of the buffer) seems pretty useless, I can use them
for something else. It seems useful to have keys to scroll by a few
lines at a time, so I'm using them for that: this will be a bit easier
than needing to hold down meta with page up / down.