Better keybindings for hyper, super and scrolling
- For scrolling by 3, use C/M-h bindings that mimic C/M-v. This was a big
motivation of this change: to give more ergonomic bindings for scrolling
by 3.

- For hyper, use C-n, since C-h and M-h are now taken. C-n is slightly
harder, but not too bad. I slightly prefer having hyper on a C or M
binding, as opposed to on a hard-to-reach thumb thing, because I
typically do hyper followed by a single key. So if it's on a
hard-to-reach thumb thing, I would want it on both hands (like I used to
have). Having it on a C or M binding seems a bit better. In retrospect,
I may move it to an easier M binding, like M-u, M-i, M-j or M-c.

- For super, use <end>. My super bindings involve multiple keys (e.g.,
super-p for prelude things, super-o for org things, etc.). So it works
okay to have it on a single, hard-to-reach thumb key; and I was finding
it too confusing to have it on a C or M binding (it involved too many
letters that were hard for me to get under my fingers).