Some customizations for storing org links that may not be important
I have had these changes uncommitted for at least a few weeks. I'm not
sure if they are still going to be important in my workflow, but I'm
committing them in case I want them.

The customization of the org-id-locations-file is useful if I decide to
use org-id, though currently I'm thinking I probably will *not* use
that.

my-org-store-link-to-file could be useful for some purposes (I think I
originally put it in for the sake of having links from my _projects.org
doc to files, though I'm not positive about that). I think that the
let-binding of org-id-link-to-org-use-id is unnecessary now that I'm not
using org-id by default, but it doesn't hurt anything either.