Swap colors of outline-5 and outline-7
I was finding outline-3 (firebrick4) and outline-5 (VioletRed4) too
close, so it was hard to tell visually which was which. I find sienna4
to be more visually distinct from those, so I'm putting that in the
middle; so now it's outline-3 and outline-7 which are visually similar,
and they are far enough apart in the outline structure that this should
be okay.

The main thing I don't like about this change is that it puts a somewhat
lighter color (sienna4) higher up in the outline hierarchy.