Easier bindings for next-logical-line and previous-logical-line
A-p in particular is difficult.

I'm not sure whether to stick with these (A-M-down/up) or use
A-down/up (removing or replacing the current bindings of A-down/up for
moving by 3 lines at a time, which I'm not finding very useful): I am on
the fence about whether A-M-down/up is easier/harder or more/less
ergonomic than A-down/up.