Use leading dots instead of a space in org-ellipsis
Motivation: I discovered that the leading space seems responsible for
the issue where, if the text of a headline runs right up to the right
margin, then the ellipsis character isn't shown (instead there is a
blank line after the headline). This issue makes it easy to miss the
fact that there is some hidden text.

It seems like using a non-space character instead solves this issue. I
experimented with a bunch of single, double and triple characters (e.g.,
characters from this set, and maybe one or two others: (`'.,:-)). I like
two dots the best. (One dot is okay, but looks like there's a period at
the end of the headline, which can look kind of funny sometimes.) This
seems approximately equally visible to having a leading space -- if
anything, it is maybe slightly less visible for done headlines with the
same color as the ellipsis and slightly more visible for other
headlines, both of which are good -- but seems to solve the above issue.